+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 
 - Make the plottable data class threadsafe ([#1588](../../pull/1588))
+- Speed up getting plottable data from items ([#1589](../../pull/1589))
 
 ## 1.29.4